+++ mrview/validator_tags.go
@@ -9,6 +9,7 @@ var (
 	regexpArticle           = regexp.MustCompile(`^[a-zA-Z0-9][a-zA-Z0-9_.+-]*[a-zA-Z0-9]$`)
 	regexpDoubleSize        = regexp.MustCompile(`^[0-9]+x[0-9]+$`)
 	regexpRewriteName       = regexp.MustCompile(`^[a-z][a-z0-9-]*[a-z0-9]$`)
+	regexpTripleSize        = regexp.MustCompile(`^[0-9]+x[0-9]+x[0-9]+$`)
 	regexpVariable          = regexp.MustCompile(`^[a-zA-Z][a-zA-Z0-9]+$`)
 )
 
@@ -25,6 +26,11 @@ func ValidateArticle(value string) bool {
 // ValidateDoubleSize - comment func.
 func ValidateDoubleSize(value string) bool {
 	return regexpDoubleSize.MatchString(value)
+}
+
+// ValidateTripleSize - comment func.
+func ValidateTripleSize(value string) bool {
+	return regexpTripleSize.MatchString(value)
 }
 
 // ValidateRewriteName - comment func.